requirements.txt
apache-airflow==2.7.2
psycopg2-binary
requests
